DeauthAttack/requirements.txt


#this code is developed using python 3.11.2
#You must install the dependencies and libraries necessary for the code to run smoothly and without errors.
#for installing ncpcap,libpcap and winpcap executables for windows

scapy===2.5.0 
colorama==0.4.4
python-nmap==0.7.1
subprocess32==3.5.4
struct==1.0.0
binascii==2.0.0
socket==0.0.1
shlex==0.0.1
python-time==0.3.0
argparse==1.4.0
pprint==0.4.0